+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Renamed `x1` and `y1` to `x` and `y`.
 - Renamed content `societyid` to `societyId`
 - Renamed content `code` to `eventcode`
+- Renamed cmd line option `--dummyValues` to `--exampleValues`
 
 ### Removed